Revert "Hotfix - BillingEntry2 & PacketLossSustained "
