+++ cmd/billing/dev.env
@@ -2,9 +2,6 @@ ENV=dev
 BACKEND_LOG_LEVEL=debug
 PORT=80
 GOOGLE_PROJECT_ID=network-next-v3-dev
-FEATURE_BILLING=true
-BILLING_TOPIC_NAME=billing
-BILLING_SUBSCRIPTION_NAME=billing
 GOOGLE_BIGQUERY_DATASET_BILLING=dev
 GOOGLE_BIGQUERY_TABLE_BILLING=billing
 GOOGLE_BIGQUERY_BATCH_SIZE=10
@@ -14,7 +11,3 @@ ENABLE_STACKDRIVER_METRICS=true
 ENABLE_STACKDRIVER_PROFILER=true
 FEATURE_ENABLE_PPROF=false
 NUM_RECEIVE_GOROUTINES=10
-FEATURE_BILLING2=true
-FEATURE_BILLING2_GOOGLE_BIGQUERY_TABLE_BILLING=billing2
-FEATURE_BILLING2_TOPIC_NAME=billing2
-FEATURE_BILLING2_SUBSCRIPTION_NAME=billing2
