+++ cmd/billing/prod.env
@@ -2,9 +2,6 @@ ENV=prod
 BACKEND_LOG_LEVEL=error
 PORT=80
 GOOGLE_PROJECT_ID=network-next-v3-prod
-FEATURE_BILLING=true
-BILLING_TOPIC_NAME=billing
-BILLING_SUBSCRIPTION_NAME=billing
 GOOGLE_BIGQUERY_DATASET_BILLING=prod
 GOOGLE_BIGQUERY_TABLE_BILLING=billing
 GOOGLE_BIGQUERY_BATCH_SIZE=1000
@@ -14,7 +11,3 @@ ENABLE_STACKDRIVER_METRICS=true
 ENABLE_STACKDRIVER_PROFILER=true
 FEATURE_ENABLE_PPROF=false
 NUM_RECEIVE_GOROUTINES=30
-FEATURE_BILLING2=true
-FEATURE_BILLING2_GOOGLE_BIGQUERY_TABLE_BILLING=billing2
-FEATURE_BILLING2_TOPIC_NAME=billing2
-FEATURE_BILLING2_SUBSCRIPTION_NAME=billing2
