--- cmd/billing/prod_debug.env
@@ -1,20 +1,0 @@
-ENV=prod
-BACKEND_LOG_LEVEL=error
-PORT=80
-GOOGLE_PROJECT_ID=network-next-v3-prod
-FEATURE_BILLING=false
-BILLING_TOPIC_NAME=billing
-BILLING_SUBSCRIPTION_NAME=billing
-GOOGLE_BIGQUERY_DATASET_BILLING=prod
-GOOGLE_BIGQUERY_TABLE_BILLING=billing
-GOOGLE_BIGQUERY_BATCH_SIZE=1000
-GOOGLE_STACKDRIVER_METRICS_WRITE_INTERVAL=1m
-ENABLE_STACKDRIVER_LOGGING=false
-ENABLE_STACKDRIVER_METRICS=true
-ENABLE_STACKDRIVER_PROFILER=true
-FEATURE_ENABLE_PPROF=false
-NUM_RECEIVE_GOROUTINES=30
-FEATURE_BILLING2=true
-FEATURE_BILLING2_GOOGLE_BIGQUERY_TABLE_BILLING=billing2
-FEATURE_BILLING2_TOPIC_NAME=billing2
-FEATURE_BILLING2_SUBSCRIPTION_NAME=billing2
