+++ cmd/billing/staging.env
@@ -2,9 +2,6 @@ ENV=staging
 BACKEND_LOG_LEVEL=none
 PORT=80
 GOOGLE_PROJECT_ID=network-next-v3-staging
-FEATURE_BILLING=true
-BILLING_TOPIC_NAME=billing
-BILLING_SUBSCRIPTION_NAME=billing
 GOOGLE_BIGQUERY_DATASET_BILLING=staging
 GOOGLE_BIGQUERY_TABLE_BILLING=billing
 GOOGLE_BIGQUERY_BATCH_SIZE=1000
@@ -14,7 +11,3 @@ ENABLE_STACKDRIVER_METRICS=true
 ENABLE_STACKDRIVER_PROFILER=true
 FEATURE_ENABLE_PPROF=false
 NUM_RECEIVE_GOROUTINES=30
-FEATURE_BILLING2=false
-FEATURE_BILLING2_GOOGLE_BIGQUERY_TABLE_BILLING=billing2
-FEATURE_BILLING2_TOPIC_NAME=billing2
-FEATURE_BILLING2_SUBSCRIPTION_NAME=billing2
