+++ cmd/server_backend/prod.env
@@ -15,11 +15,9 @@ MAXMIND_ISP_DB_FILE=/app/GeoIP2-ISP.mmdb
 MAXMIND_SYNC_DB_INTERVAL=1m
 ROUTE_MATRIX_URI=http://35.227.196.44/route_matrix
 ROUTE_MATRIX_SYNC_INTERVAL=1s
-FEATURE_BILLING=true
 BILLING_CLIENT_COUNT=1
 BILLING_BATCHED_MESSAGE_COUNT=100
 BILLING_BATCHED_MESSAGE_MIN_BYTES=1024
-BILLING_TOPIC_NAME=billing
 PORTAL_CRUNCHER_HOSTS=tcp://10.128.0.130:5555,tcp://10.128.0.141:5555,tcp://10.128.0.150:5555,tcp://10.128.0.165:5555
 POST_SESSION_THREAD_COUNT=1000
 POST_SESSION_BUFFER_SIZE=1000000
@@ -39,5 +37,3 @@ FEATURE_VANITY_METRIC_POST_SEND_BUFFER_SIZE=1000000
 FEATURE_VANITY_METRIC_POST_MAX_RETRIES=10
 FEATURE_ENABLE_PPROF=false
 MATRIX_STALE_DURATION=20s
-FEATURE_BILLING2=true
-FEATURE_BILLING2_TOPIC_NAME=billing2
