+++ cmd/server_backend/staging_debug.env
@@ -15,11 +15,9 @@ MAXMIND_ISP_DB_FILE=/app/GeoIP2-ISP.mmdb
 MAXMIND_SYNC_DB_INTERVAL=1m
 ROUTE_MATRIX_URI=http://35.190.44.124/route_matrix
 ROUTE_MATRIX_SYNC_INTERVAL=1s
-FEATURE_BILLING=true
 BILLING_CLIENT_COUNT=1
 BILLING_BATCHED_MESSAGE_COUNT=100
 BILLING_BATCHED_MESSAGE_MIN_BYTES=1024
-BILLING_TOPIC_NAME=billing
 PORTAL_CRUNCHER_HOSTS=tcp://10.128.0.13:5555,tcp://10.128.0.49:5555,tcp://10.128.0.89:5555,tcp://10.128.0.143:5555
 POST_SESSION_THREAD_COUNT=1000
 POST_SESSION_BUFFER_SIZE=5000000
@@ -41,5 +39,3 @@ FEATURE_ENABLE_PPROF=false
 NEXT_DEBUG_LOGS=1
 NEXT_DEBUG=true
 MATRIX_STALE_DURATION=20s
-FEATURE_BILLING2=false
-FEATURE_BILLING2_TOPIC_NAME=billing2
