+++ deploy/build-artifacts.sh
@@ -44,11 +44,6 @@ build-artifacts() {
 		cp ${DIR}/../cmd/relay_backend/${ENV}_debug.env ${DIST_DIR}/artifact/${SERVICE}/app.env
 		cp ${DIR}/${SYSTEMD_SERVICE_FILE} ${DIST_DIR}/artifact/${SERVICE}/${SYSTEMD_SERVICE_FILE}
 		cd ${DIST_DIR}/artifact/${SERVICE} && tar -zcf ../../${SERVICE}.${ENV}.tar.gz app app.env ${SYSTEMD_SERVICE_FILE} && cd ../..
-  elif [ "$SERVICE" = "debug_billing" ]; then
-    cp ${DIST_DIR}/billing ${DIST_DIR}/artifact/${SERVICE}/app
-    cp ${DIR}/../cmd/billing/${ENV}_debug.env ${DIST_DIR}/artifact/${SERVICE}/app.env
-    cp ${DIR}/${SYSTEMD_SERVICE_FILE} ${DIST_DIR}/artifact/${SERVICE}/${SYSTEMD_SERVICE_FILE}
-    cd ${DIST_DIR}/artifact/${SERVICE} && tar -zcf ../../${SERVICE}.${ENV}.tar.gz app app.env ${SYSTEMD_SERVICE_FILE} && cd ../..
 	else
 		cp ${DIST_DIR}/${SERVICE} ${DIST_DIR}/artifact/${SERVICE}/app
 		cp ${DIR}/../cmd/${SERVICE}/${ENV}.env ${DIST_DIR}/artifact/${SERVICE}/app.env
