+++ modules/billing/billing_entry.go
@@ -2,9 +2,6 @@ package billing
 
 import (
 	"fmt"
-	"math"
-
-	"cloud.google.com/go/bigquery"
 
 	"github.com/networknext/backend/modules/encoding"
 )
@@ -718,1212 +715,4 @@ func ReadBillingEntry(entry *BillingEntry, data []byte) bool {
 	}
 
 	return true
-}
-
-// Validate a billing entry. Returns true if the billing entry is valid, false if invalid.
-func (entry *BillingEntry) Validate() bool {
-
-	if entry.Timestamp < 0 {
-		fmt.Printf("invalid timestamp\n")
-		return false
-	}
-
-	if entry.BuyerID == 0 {
-		fmt.Printf("invalid buyer id\n")
-		return false
-	}
-
-	if entry.SessionID == 0 {
-		fmt.Printf("invalid session id\n")
-		return false
-	}
-
-	// IMPORTANT: Logic inverted because comparing a NaN float value always returns false
-	if !(entry.Latitude >= -90.0 && entry.Latitude <= +90.0) {
-		fmt.Printf("invalid latitude\n")
-		return false
-	}
-
-	if !(entry.Longitude >= -180.0 && entry.Longitude <= +180.0) {
-		fmt.Printf("invalid longitude\n")
-		return false
-	}
-
-	// IMPORTANT: You must update this check if you ever add a new connection type in the SDK
-	if entry.ConnectionType < 0 || entry.ConnectionType > 3 {
-		fmt.Printf("invalid connection type\n")
-		return false
-	}
-
-	// IMPORTANT: You must update this check when you add new platforms to the SDK
-	if entry.PlatformType < 0 || entry.ConnectionType > 10 {
-		fmt.Printf("invalid connection type\n")
-		return false
-	}
-
-	if entry.RouteDiversity < 0 || entry.RouteDiversity > 32 {
-		fmt.Printf("invalid route diversity\n")
-		return false
-	}
-
-	if !(entry.DirectRTT >= 0.0 && entry.DirectRTT <= 10000.0) {
-		fmt.Printf("invalid direct rtt\n")
-		return false
-	}
-
-	if !(entry.DirectJitter >= 0.0 && entry.DirectJitter <= 10000.0) {
-		fmt.Printf("invalid direct jitter\n")
-		return false
-	}
-
-	if !(entry.DirectPacketLoss >= 0.0 && entry.DirectPacketLoss <= 100.0) {
-		fmt.Printf("invalid direct packet loss\n")
-		return false
-	}
-
-	if !(entry.PacketLoss >= 0.0 && entry.PacketLoss <= 100.0) {
-		if entry.PacketLoss > 100.0 {
-			fmt.Printf("PacketLoss %v > 100.0. Clamping to 100.0\n%+v\n", entry.PacketLoss, entry)
-			entry.PacketLoss = 100.0
-		} else {
-			fmt.Printf("invalid packet loss\n")
-			return false
-		}
-	}
-
-	if !(entry.JitterClientToServer >= 0.0 && entry.JitterClientToServer <= 1000.0) {
-		if entry.JitterClientToServer > 1000.0 {
-			fmt.Printf("JitterClientToServer %v > 1000.0. Clamping to 1000.0\n%+v\n", entry.JitterClientToServer, entry)
-			entry.JitterClientToServer = 1000.0
-		} else {
-			fmt.Printf("invalid jitter client to server\n")
-			return false
-		}
-	}
-
-	if !(entry.JitterServerToClient >= 0.0 && entry.JitterServerToClient <= 1000.0) {
-		if entry.JitterServerToClient > 1000.0 {
-			fmt.Printf("JitterServerToClient %v > 1000.0. Clamping to 1000.0.\n%+v\n", entry.JitterServerToClient, entry)
-			entry.JitterServerToClient = 1000.0
-		} else {
-			fmt.Printf("invalid jitter server to client\n")
-			return false
-		}
-	}
-
-	if entry.NumTags < 0 || entry.NumTags > 8 {
-		fmt.Printf("invalid num tags\n")
-		return false
-	}
-
-	if entry.Next {
-
-		if !(entry.NextRTT >= 0.0 && entry.NextRTT <= 10000.0) {
-			fmt.Printf("invalid next rtt\n")
-			return false
-		}
-
-		if !(entry.NextJitter >= 0.0 && entry.NextJitter <= 10000.0) {
-			fmt.Printf("invalid next jitter\n")
-			return false
-		}
-
-		if !(entry.NextPacketLoss >= 0.0 && entry.NextPacketLoss <= 100.0) {
-			fmt.Printf("invalid next packet loss\n")
-			return false
-		}
-
-		if !(entry.PredictedNextRTT >= 0.0 && entry.PredictedNextRTT <= 10000.0) {
-			fmt.Printf("invalid predicted next rtt\n")
-			return false
-		}
-
-		if entry.NumNearRelays < 0 || entry.NumNearRelays > 32 {
-			fmt.Printf("invalid num near relays\n")
-			return false
-		}
-
-		if entry.NumNearRelays > 0 {
-
-			for i := 0; i < int(entry.NumNearRelays); i++ {
-
-				if entry.NearRelayIDs[i] == 0 {
-					// Log this but do not return false
-					// TODO: investigate why nearRelayID is 0
-					fmt.Printf("NearRelayIDs[%d] is 0.\n%+v\n", i, entry)
-				}
-
-				if !(entry.NearRelayRTTs[i] >= 0.0 && entry.NearRelayRTTs[i] <= 255.0) {
-					fmt.Printf("invalid near relay rtt\n")
-					return false
-				}
-
-				if !(entry.NearRelayJitters[i] >= 0.0 && entry.NearRelayJitters[i] <= 255.0) {
-					fmt.Printf("invalid near relay jitter\n")
-					return false
-				}
-
-				if !(entry.NearRelayPacketLosses[i] >= 0.0 && entry.NearRelayPacketLosses[i] <= 100.0) {
-					fmt.Printf("invalid near relay packet loss\n")
-					return false
-				}
-			}
-		}
-	}
-
-	return true
-}
-
-// Checks all floating point numbers in the BillingEntry for NaN and +-Inf and forces them to 0
-func (entry *BillingEntry) CheckNaNOrInf() (bool, []string) {
-	var nanOrInfExists bool
-	var nanOrInfFields []string
-
-	if math.IsNaN(float64(entry.NextRTT)) || math.IsInf(float64(entry.NextRTT), 0) {
-		nanOrInfFields = append(nanOrInfFields, "NextRTT")
-		nanOrInfExists = true
-		entry.NextRTT = float32(0)
-	}
-
-	if math.IsNaN(float64(entry.NextJitter)) || math.IsInf(float64(entry.NextJitter), 0) {
-		nanOrInfFields = append(nanOrInfFields, "NextJitter")
-		nanOrInfExists = true
-		entry.NextJitter = float32(0)
-	}
-
-	if math.IsNaN(float64(entry.NextPacketLoss)) || math.IsInf(float64(entry.NextPacketLoss), 0) {
-		nanOrInfFields = append(nanOrInfFields, "NextPacketLoss")
-		nanOrInfExists = true
-		entry.NextPacketLoss = float32(0)
-	}
-
-	if math.IsNaN(float64(entry.Latitude)) || math.IsInf(float64(entry.Latitude), 0) {
-		nanOrInfFields = append(nanOrInfFields, "Latitude")
-		nanOrInfExists = true
-		entry.Latitude = float32(0)
-	}
-
-	if math.IsNaN(float64(entry.Longitude)) || math.IsInf(float64(entry.Longitude), 0) {
-		nanOrInfFields = append(nanOrInfFields, "Longitude")
-		nanOrInfExists = true
-		entry.Longitude = float32(0)
-	}
-
-	if math.IsNaN(float64(entry.PacketLoss)) || math.IsInf(float64(entry.PacketLoss), 0) {
-		nanOrInfFields = append(nanOrInfFields, "PacketLoss")
-		nanOrInfExists = true
-		entry.PacketLoss = float32(0)
-	}
-
-	if math.IsNaN(float64(entry.PredictedNextRTT)) || math.IsInf(float64(entry.PredictedNextRTT), 0) {
-		nanOrInfFields = append(nanOrInfFields, "PredictedNextRTT")
-		nanOrInfExists = true
-		entry.PredictedNextRTT = float32(0)
-	}
-
-	if math.IsNaN(float64(entry.NearRelayRTT)) || math.IsInf(float64(entry.NearRelayRTT), 0) {
-		nanOrInfFields = append(nanOrInfFields, "NearRelayRTT")
-		nanOrInfExists = true
-		entry.NearRelayRTT = float32(0)
-	}
-
-	if math.IsNaN(float64(entry.JitterClientToServer)) || math.IsInf(float64(entry.JitterClientToServer), 0) {
-		nanOrInfFields = append(nanOrInfFields, "JitterClientToServer")
-		nanOrInfExists = true
-		entry.JitterClientToServer = float32(0)
-	}
-
-	if math.IsNaN(float64(entry.JitterServerToClient)) || math.IsInf(float64(entry.JitterServerToClient), 0) {
-		nanOrInfFields = append(nanOrInfFields, "JitterServerToClient")
-		nanOrInfExists = true
-		entry.JitterServerToClient = float32(0)
-	}
-
-	if entry.NumNearRelays > 0 {
-		for i := 0; i < int(entry.NumNearRelays); i++ {
-			if math.IsNaN(float64(entry.NearRelayRTTs[i])) || math.IsInf(float64(entry.NearRelayRTTs[i]), 0) {
-				nanOrInfFields = append(nanOrInfFields, fmt.Sprintf("NearRelayRTTs[%d]", i))
-				nanOrInfExists = true
-				entry.NearRelayRTTs[i] = float32(0)
-			}
-
-			if math.IsNaN(float64(entry.NearRelayJitters[i])) || math.IsInf(float64(entry.NearRelayJitters[i]), 0) {
-				nanOrInfFields = append(nanOrInfFields, fmt.Sprintf("NearRelayJitters[%d]", i))
-				nanOrInfExists = true
-				entry.NearRelayJitters[i] = float32(0)
-			}
-
-			if math.IsNaN(float64(entry.NearRelayPacketLosses[i])) || math.IsInf(float64(entry.NearRelayPacketLosses[i]), 0) {
-				nanOrInfFields = append(nanOrInfFields, fmt.Sprintf("NearRelayPacketLosses[%d]", i))
-				nanOrInfExists = true
-				entry.NearRelayPacketLosses[i] = float32(0)
-			}
-		}
-	}
-
-	return nanOrInfExists, nanOrInfFields
-}
-
-// Implements the bigquery.ValueSaver interface for a billing entry so it can be used in Put()
-func (entry *BillingEntry) Save() (map[string]bigquery.Value, string, error) {
-
-	e := make(map[string]bigquery.Value)
-
-	e["timestamp"] = int(entry.Timestamp)
-
-	e["buyerId"] = int(entry.BuyerID)
-	e["sessionId"] = int(entry.SessionID)
-	e["datacenterID"] = int(entry.DatacenterID)
-	e["userHash"] = int(entry.UserHash)
-	e["latitude"] = entry.Latitude
-	e["longitude"] = entry.Longitude
-	e["isp"] = entry.ISP
-	e["connectionType"] = int(entry.ConnectionType)
-	e["platformType"] = int(entry.PlatformType)
-	e["sdkVersion"] = entry.SDKVersion
-
-	e["sliceNumber"] = int(entry.SliceNumber)
-
-	e["flagged"] = entry.Flagged
-	e["fallbackToDirect"] = entry.FallbackToDirect
-	e["multipathVetoed"] = entry.MultipathVetoed
-	e["abTest"] = entry.ABTest
-	e["committed"] = entry.Committed
-	e["multipath"] = entry.Multipath
-	e["rttReduction"] = entry.RTTReduction
-	e["packetLossReduction"] = entry.PacketLossReduction
-	e["relayWentAway"] = entry.RelayWentAway
-	e["routeLost"] = entry.RouteLost
-	e["mispredicted"] = entry.Mispredicted
-	e["vetoed"] = entry.Vetoed
-	e["latencyWorse"] = entry.LatencyWorse
-	e["noRoute"] = entry.NoRoute
-	e["nextLatencyTooHigh"] = entry.NextLatencyTooHigh
-	e["routeChanged"] = entry.RouteChanged
-	e["commitVeto"] = entry.CommitVeto
-
-	if entry.Pro {
-		e["pro"] = entry.Pro
-	}
-
-	if entry.BuyerNotLive {
-		e["buyerNotLive"] = entry.BuyerNotLive
-	}
-
-	if entry.UnknownDatacenter {
-		e["unknownDatacenter"] = entry.UnknownDatacenter
-	}
-
-	if entry.DatacenterNotEnabled {
-		e["datacenterNotEnabled"] = entry.DatacenterNotEnabled
-	}
-
-	if entry.StaleRouteMatrix {
-		e["staleRouteMatrix"] = entry.StaleRouteMatrix
-	}
-
-	if entry.RouteDiversity > 0 {
-		e["routeDiversity"] = entry.RouteDiversity
-	}
-
-	if entry.LackOfDiversity {
-		e["lackOfDiversity"] = entry.LackOfDiversity
-	}
-
-	if entry.MultipathRestricted {
-		e["multipathRestricted"] = entry.MultipathRestricted
-	}
-
-	e["directRTT"] = entry.DirectRTT
-	e["directJitter"] = entry.DirectJitter
-	e["directPacketLoss"] = entry.DirectPacketLoss
-
-	if entry.ClientToServerPacketsSent > 0 {
-		e["clientToServerPacketsSent"] = int(entry.ClientToServerPacketsSent)
-	}
-
-	if entry.ServerToClientPacketsSent > 0 {
-		e["serverToClientPacketsSent"] = int(entry.ServerToClientPacketsSent)
-	}
-
-	if entry.ClientToServerPacketsLost > 0 {
-		e["clientToServerPacketsLost"] = int(entry.ClientToServerPacketsLost)
-	}
-
-	if entry.ServerToClientPacketsLost > 0 {
-		e["serverToClientPacketsLost"] = int(entry.ServerToClientPacketsLost)
-	}
-
-	// IMPORTANT: This is derived from *PacketsSent and *PacketsLost, and is valid for both next and direct
-	if entry.PacketLoss > 0.0 {
-		e["packetLoss"] = entry.PacketLoss
-	}
-
-	if entry.PacketsOutOfOrderClientToServer != 0 {
-		e["packetsOutOfOrderClientToServer"] = int(entry.PacketsOutOfOrderClientToServer)
-	}
-
-	if entry.PacketsOutOfOrderServerToClient != 0 {
-		e["packetsOutOfOrderServerToClient"] = int(entry.PacketsOutOfOrderServerToClient)
-	}
-
-	if entry.JitterClientToServer != 0 {
-		e["jitterClientToServer"] = entry.JitterClientToServer
-	}
-
-	if entry.JitterServerToClient != 0 {
-		e["jitterServerToClient"] = entry.JitterServerToClient
-	}
-
-	if entry.UseDebug && entry.Debug != "" {
-		e["debug"] = entry.Debug
-	}
-
-	if entry.ClientFlags != 0 {
-		e["clientFlags"] = int(entry.ClientFlags)
-	}
-
-	if entry.UserFlags != 0 {
-		e["userFlags"] = int(entry.UserFlags)
-	}
-
-	if entry.NumTags > 0 {
-		tags := make([]bigquery.Value, entry.NumTags)
-		for i := 0; i < int(entry.NumTags); i++ {
-			tags[i] = int(entry.Tags[i])
-		}
-		e["tags"] = tags
-	}
-
-	if entry.Next {
-
-		e["next"] = entry.Next
-
-		e["nextRTT"] = entry.NextRTT
-		e["nextJitter"] = entry.NextJitter
-		e["nextPacketLoss"] = entry.NextPacketLoss
-
-		e["totalPrice"] = int(entry.TotalPrice)
-
-		e["nextBytesUp"] = int(entry.NextBytesUp)
-		e["nextBytesDown"] = int(entry.NextBytesDown)
-		e["envelopeBytesUp"] = int(entry.EnvelopeBytesUp)
-		e["envelopeBytesDown"] = int(entry.EnvelopeBytesDown)
-
-		if entry.PredictedNextRTT > 0.0 {
-			e["predictedNextRTT"] = entry.PredictedNextRTT
-		}
-
-		if entry.NearRelayRTT != 0 {
-			e["nearRelayRTT"] = entry.NearRelayRTT
-		}
-
-		nextRelays := make([]bigquery.Value, entry.NumNextRelays)
-		nextRelaysPrice := make([]bigquery.Value, entry.NumNextRelays)
-
-		for i := 0; i < int(entry.NumNextRelays); i++ {
-			nextRelays[i] = int(entry.NextRelays[i])
-			nextRelaysPrice[i] = int(entry.NextRelaysPrice[i])
-		}
-
-		e["nextRelays"] = nextRelays
-		e["nextRelaysPrice"] = nextRelaysPrice
-
-		if entry.NumNearRelays > 0 && entry.UseDebug {
-			// IMPORTANT: Only write this data if debug is on because it is very large
-
-			e["numNearRelays"] = int(entry.NumNearRelays)
-
-			nearRelayIDs := make([]bigquery.Value, entry.NumNearRelays)
-			nearRelayRTTs := make([]bigquery.Value, entry.NumNearRelays)
-			nearRelayJitters := make([]bigquery.Value, entry.NumNearRelays)
-			nearRelayPacketLosses := make([]bigquery.Value, entry.NumNearRelays)
-
-			for i := 0; i < int(entry.NumNearRelays); i++ {
-				nearRelayIDs[i] = int(entry.NearRelayIDs[i])
-				nearRelayRTTs[i] = entry.NearRelayRTTs[i]
-				nearRelayJitters[i] = entry.NearRelayJitters[i]
-				nearRelayPacketLosses[i] = entry.NearRelayPacketLosses[i]
-			}
-
-			e["nearRelayIDs"] = nearRelayIDs
-			e["nearRelayRTTs"] = nearRelayRTTs
-			e["nearRelayJitters"] = nearRelayJitters
-			e["nearRelayPacketLosses"] = nearRelayPacketLosses
-		}
-	}
-
-	// todo: this is deprecated. we don't really need this anymore. should
-	// be made nullable in the schema and we just stop writing this.
-	e["initial"] = entry.Initial
-
-	// todo: this is deprecated and should be made nullable in the schema
-	// at this point we should stop writing this
-	e["routeDecision"] = int(entry.RouteDecision)
-
-	return e, "", nil
-}
-
-// ------------------------------------------------------------------------
-
-const (
-	BillingEntryVersion2 = uint8(0)
-
-	MaxBillingEntry2Bytes = 4 + // Version
-		4 + // Timestamp
-		8 + // SessionID
-		4 + // SliceNumber
-		4 + // DirectRTT
-		4 + // DirectJitter
-		4 + // DirectPacketLoss
-		4 + // RealPacketLoss
-		4 + // RealPacketLoss_Frac
-		4 + // RealJitter
-		1 + // Next
-		1 + // Flagged
-		1 + // Summary
-		1 + // UseDebug
-		4 + BillingEntryMaxDebugLength + // Debug
-		8 + // DatacenterID
-		8 + // BuyerID
-		8 + // UserHash
-		8 + // EnvelopeBytesUp
-		8 + // EnvelopeBytesDown
-		4 + // Latitude
-		4 + // Longitude
-		4 + BillingEntryMaxISPLength + // ISP
-		4 + // ConnectionType
-		4 + // PlatformType
-		4 + BillingEntryMaxSDKVersionLength + // SDKVersion
-		4 + // NumTags
-		BillingEntryMaxTags*8 + // Tags
-		1 + // ABTest
-		1 + // Pro
-		8 + // ClientToServerPacketsSent
-		8 + // ServerToClientPacketsSent
-		8 + // ClientToServerPacketsLost
-		8 + // ServerToClientPacketsLost
-		8 + // ClientToServerPacketsOutOfOrder
-		8 + // ServerToClientPacketsOutOfOrder
-		4 + // NumNearRelays
-		BillingEntryMaxNearRelays*8 + // NearRelayIDs
-		BillingEntryMaxNearRelays*4 + // NearRelayRTTs
-		BillingEntryMaxNearRelays*4 + // NearRelayJitters
-		BillingEntryMaxNearRelays*4 + // NearRelayPacketLosses
-		4 + // NextRTT
-		4 + // NextJitter
-		4 + // NextPacketLoss
-		4 + // PredictedNextRTT
-		4 + // NearRelayRTT
-		4 + // NumNextRelays
-		BillingEntryMaxRelays*8 + // NextRelays
-		BillingEntryMaxRelays*8 + // NextRelayPrice
-		8 + // TotalPrice
-		4 + // RouteDiversity
-		1 + // Uncommitted
-		1 + // Multipath
-		1 + // RTTReduction
-		1 + // PacketLossReduction
-		1 + // RouteChanged
-		1 + // FallbackToDirect
-		1 + // MultipathVetoed
-		1 + // Mispredicted
-		1 + // Vetoed
-		1 + // LatencyWorse
-		1 + // NoRoute
-		1 + // NextLatencyTooHigh
-		1 + // CommitVeto
-		1 + // UnknownDatacenter
-		1 + // DatacenterNotEnabled
-		1 + // BuyerNotLive
-		1 + // StaleRouteMatrix
-		2 // 2 extra bytes to let size be divisible by 4
-)
-
-type BillingEntry2 struct {
-
-	// always
-
-	Version             uint32
-	Timestamp           uint32
-	SessionID           uint64
-	SliceNumber         uint32
-	DirectRTT           int32
-	DirectJitter        int32
-	DirectPacketLoss    int32
-	RealPacketLoss      int32
-	RealPacketLoss_Frac uint32
-	RealJitter          uint32
-	Next                bool
-	Flagged             bool
-	Summary             bool
-	UseDebug            bool
-	Debug               string
-
-	// first slice only
-
-	DatacenterID      uint64
-	BuyerID           uint64
-	UserHash          uint64
-	EnvelopeBytesUp   uint64
-	EnvelopeBytesDown uint64
-	Latitude          float32
-	Longitude         float32
-	ISP               string
-	ConnectionType    int32
-	PlatformType      int32
-	SDKVersion        string
-	NumTags           int32
-	Tags              [BillingEntryMaxTags]uint64
-	ABTest            bool
-	Pro               bool
-
-	// summary slice only
-
-	ClientToServerPacketsSent       uint64
-	ServerToClientPacketsSent       uint64
-	ClientToServerPacketsLost       uint64
-	ServerToClientPacketsLost       uint64
-	ClientToServerPacketsOutOfOrder uint64
-	ServerToClientPacketsOutOfOrder uint64
-	NumNearRelays                   int32
-	NearRelayIDs                    [BillingEntryMaxNearRelays]uint64
-	NearRelayRTTs                   [BillingEntryMaxNearRelays]int32
-	NearRelayJitters                [BillingEntryMaxNearRelays]int32
-	NearRelayPacketLosses           [BillingEntryMaxNearRelays]int32
-
-	// network next only
-
-	NextRTT             int32
-	NextJitter          int32
-	NextPacketLoss      int32
-	PredictedNextRTT    int32
-	NearRelayRTT        int32
-	NumNextRelays       int32
-	NextRelays          [BillingEntryMaxRelays]uint64
-	NextRelayPrice      [BillingEntryMaxRelays]uint64
-	TotalPrice          uint64
-	RouteDiversity      int32
-	Uncommitted         bool
-	Multipath           bool
-	RTTReduction        bool
-	PacketLossReduction bool
-	RouteChanged        bool
-
-	// error state only
-
-	FallbackToDirect     bool
-	MultipathVetoed      bool
-	Mispredicted         bool
-	Vetoed               bool
-	LatencyWorse         bool
-	NoRoute              bool
-	NextLatencyTooHigh   bool
-	CommitVeto           bool
-	UnknownDatacenter    bool
-	DatacenterNotEnabled bool
-	BuyerNotLive         bool
-	StaleRouteMatrix     bool
-}
-
-func (entry *BillingEntry2) Serialize(stream encoding.Stream) error {
-
-	/*
-		1. Always
-
-		These values are serialized in every slice
-	*/
-
-	stream.SerializeBits(&entry.Version, 8)
-	stream.SerializeBits(&entry.Timestamp, 32)
-	stream.SerializeUint64(&entry.SessionID)
-
-	small := false
-	if entry.SliceNumber < 1024 {
-		small = true
-	}
-	stream.SerializeBool(&small)
-	if small {
-		stream.SerializeBits(&entry.SliceNumber, 10)
-	} else {
-		stream.SerializeBits(&entry.SliceNumber, 32)
-	}
-
-	stream.SerializeInteger(&entry.DirectRTT, 0, 1023)
-	stream.SerializeInteger(&entry.DirectJitter, 0, 255)
-	stream.SerializeInteger(&entry.DirectPacketLoss, 0, 100)
-
-	stream.SerializeInteger(&entry.RealPacketLoss, 0, 100)
-	stream.SerializeBits(&entry.RealPacketLoss_Frac, 8)
-	stream.SerializeBits(&entry.RealJitter, 8)
-
-	stream.SerializeBool(&entry.Next)
-	stream.SerializeBool(&entry.Flagged)
-	stream.SerializeBool(&entry.Summary)
-
-	stream.SerializeBool(&entry.UseDebug)
-	stream.SerializeString(&entry.Debug, BillingEntryMaxDebugLength)
-
-	/*
-		2. First slice only
-
-		These values are serialized only for slice 0.
-	*/
-
-	if entry.SliceNumber == 0 {
-
-		stream.SerializeUint64(&entry.DatacenterID)
-		stream.SerializeUint64(&entry.BuyerID)
-		stream.SerializeUint64(&entry.UserHash)
-		stream.SerializeUint64(&entry.EnvelopeBytesUp)
-		stream.SerializeUint64(&entry.EnvelopeBytesDown)
-		stream.SerializeFloat32(&entry.Latitude)
-		stream.SerializeFloat32(&entry.Longitude)
-		stream.SerializeString(&entry.ISP, BillingEntryMaxISPLength)
-		stream.SerializeInteger(&entry.ConnectionType, 0, 3) // todo: constant
-		stream.SerializeInteger(&entry.PlatformType, 0, 10)  // todo: constant
-		stream.SerializeString(&entry.SDKVersion, 12)        // todo: constant
-		stream.SerializeInteger(&entry.NumTags, 0, BillingEntryMaxTags)
-		for i := 0; i < int(entry.NumTags); i++ {
-			stream.SerializeUint64(&entry.Tags[i])
-		}
-		stream.SerializeBool(&entry.ABTest)
-		stream.SerializeBool(&entry.Pro)
-
-	}
-
-	/*
-		3. Summary slice only
-
-		These values are serialized only for the summary slice (at the end of the session)
-	*/
-
-	if entry.Summary {
-
-		stream.SerializeUint64(&entry.ClientToServerPacketsSent)
-		stream.SerializeUint64(&entry.ServerToClientPacketsSent)
-		stream.SerializeUint64(&entry.ClientToServerPacketsLost)
-		stream.SerializeUint64(&entry.ServerToClientPacketsLost)
-		stream.SerializeUint64(&entry.ClientToServerPacketsOutOfOrder)
-		stream.SerializeUint64(&entry.ServerToClientPacketsOutOfOrder)
-		stream.SerializeInteger(&entry.NumNearRelays, 0, BillingEntryMaxNearRelays)
-		for i := 0; i < int(entry.NumNearRelays); i++ {
-			stream.SerializeUint64(&entry.NearRelayIDs[i])
-			stream.SerializeInteger(&entry.NearRelayRTTs[i], 0, 255)
-			stream.SerializeInteger(&entry.NearRelayJitters[i], 0, 255)
-			stream.SerializeInteger(&entry.NearRelayPacketLosses[i], 0, 100)
-		}
-
-	}
-
-	/*
-		4. Network Next Only
-
-		These values are serialized only when a slice is on network next.
-	*/
-
-	if entry.Next {
-
-		stream.SerializeInteger(&entry.NextRTT, 0, 255)
-		stream.SerializeInteger(&entry.NextJitter, 0, 255)
-		stream.SerializeInteger(&entry.NextPacketLoss, 0, 100)
-		stream.SerializeInteger(&entry.PredictedNextRTT, 0, 255)
-		stream.SerializeInteger(&entry.NearRelayRTT, 0, 255)
-
-		stream.SerializeInteger(&entry.NumNextRelays, 0, BillingEntryMaxRelays)
-		for i := 0; i < int(entry.NumNextRelays); i++ {
-			stream.SerializeUint64(&entry.NextRelays[i])
-			stream.SerializeUint64(&entry.NextRelayPrice[i])
-		}
-
-		stream.SerializeUint64(&entry.TotalPrice)
-		stream.SerializeInteger(&entry.RouteDiversity, 0, 31)
-		stream.SerializeBool(&entry.Uncommitted)
-		stream.SerializeBool(&entry.Multipath)
-		stream.SerializeBool(&entry.RTTReduction)
-		stream.SerializeBool(&entry.PacketLossReduction)
-		stream.SerializeBool(&entry.RouteChanged)
-	}
-
-	/*
-		5. Error State Only
-
-		These values are only serialized when the session is in an error state (rare).
-	*/
-
-	errorState := false
-
-	if stream.IsWriting() {
-		errorState =
-			entry.FallbackToDirect ||
-				entry.MultipathVetoed ||
-				entry.Mispredicted ||
-				entry.Vetoed ||
-				entry.LatencyWorse ||
-				entry.NoRoute ||
-				entry.NextLatencyTooHigh ||
-				entry.CommitVeto ||
-				entry.UnknownDatacenter ||
-				entry.DatacenterNotEnabled ||
-				entry.BuyerNotLive ||
-				entry.StaleRouteMatrix
-	}
-
-	stream.SerializeBool(&errorState)
-
-	if errorState {
-
-		stream.SerializeBool(&entry.FallbackToDirect)
-		stream.SerializeBool(&entry.MultipathVetoed)
-		stream.SerializeBool(&entry.Mispredicted)
-		stream.SerializeBool(&entry.Vetoed)
-		stream.SerializeBool(&entry.LatencyWorse)
-		stream.SerializeBool(&entry.NoRoute)
-		stream.SerializeBool(&entry.NextLatencyTooHigh)
-		stream.SerializeBool(&entry.CommitVeto)
-		stream.SerializeBool(&entry.UnknownDatacenter)
-		stream.SerializeBool(&entry.DatacenterNotEnabled)
-		stream.SerializeBool(&entry.BuyerNotLive)
-		stream.SerializeBool(&entry.StaleRouteMatrix)
-
-	}
-
-	return stream.Error()
-}
-
-func WriteBillingEntry2(entry *BillingEntry2) ([]byte, error) {
-	defer func() {
-		if r := recover(); r != nil {
-			fmt.Printf("recovered from panic during BillingEntry2 packet entry write: %v\n", r)
-		}
-	}()
-
-	buffer := [MaxBillingEntry2Bytes]byte{}
-
-	ws, err := encoding.CreateWriteStream(buffer[:])
-	if err != nil {
-		return nil, err
-	}
-
-	if err := entry.Serialize(ws); err != nil {
-		return nil, err
-	}
-	ws.Flush()
-
-	return buffer[:ws.GetBytesProcessed()], nil
-}
-
-func ReadBillingEntry2(entry *BillingEntry2, data []byte) error {
-	if err := entry.Serialize(encoding.CreateReadStream(data)); err != nil {
-		return err
-	}
-	return nil
-}
-
-// Validate a billing entry 2. Returns true if the billing entry 2 is valid, false if invalid.
-func (entry *BillingEntry2) Validate() bool {
-
-	// always
-
-	if entry.Version < 0 {
-		fmt.Printf("invalid version\n")
-		return false
-	}
-
-	if entry.Timestamp < 0 {
-		fmt.Printf("invalid timestamp\n")
-		return false
-	}
-
-	if entry.SessionID == 0 {
-		fmt.Printf("invalid session id\n")
-		return false
-	}
-
-	if entry.SliceNumber < 0 {
-		fmt.Printf("invalid slice number\n")
-		return false
-	}
-
-	if entry.DirectRTT < 0 || entry.DirectRTT > 1023 {
-		fmt.Printf("invalid direct rtt\n")
-		return false
-	}
-
-	if entry.DirectJitter < 0 || entry.DirectJitter > 255 {
-		fmt.Printf("invalid direct jitter\n")
-		return false
-	}
-
-	if entry.DirectPacketLoss < 0 || entry.DirectPacketLoss > 100 {
-		fmt.Printf("invalid direct packet loss\n")
-		return false
-	}
-
-	if entry.RealPacketLoss < 0 || entry.RealPacketLoss > 100 {
-		if entry.RealPacketLoss > 100 {
-			fmt.Printf("RealPacketLoss %v > 100. Clamping to 100\n%+v\n", entry.RealPacketLoss, entry)
-			entry.RealPacketLoss = 100
-		} else {
-			fmt.Printf("invalid real packet loss\n")
-			return false
-		}
-	}
-
-	// first slice only
-
-	if entry.SliceNumber == 0 {
-
-		if entry.BuyerID == 0 {
-			fmt.Printf("invalid buyer id\n")
-			return false
-		}
-
-		// IMPORTANT: Logic inverted because comparing a NaN float value always returns false
-		if !(entry.Latitude >= -90.0 && entry.Latitude <= +90.0) {
-			fmt.Printf("invalid latitude\n")
-			return false
-		}
-
-		if !(entry.Longitude >= -180.0 && entry.Longitude <= +180.0) {
-			fmt.Printf("invalid longitude\n")
-			return false
-		}
-
-		// IMPORTANT: You must update this check if you ever add a new connection type in the SDK
-		if entry.ConnectionType < 0 || entry.ConnectionType > 3 {
-			fmt.Printf("invalid connection type\n")
-			return false
-		}
-
-		// IMPORTANT: You must update this check when you add new platforms to the SDK
-		if entry.PlatformType < 0 || entry.ConnectionType > 10 {
-			fmt.Printf("invalid connection type\n")
-			return false
-		}
-
-		if entry.NumTags < 0 || entry.NumTags > 8 {
-			fmt.Printf("invalid num tags\n")
-			return false
-		}
-	}
-
-	// summary slice only
-
-	if entry.Summary {
-
-		if entry.NumNearRelays < 0 || entry.NumNearRelays > 32 {
-			fmt.Printf("invalid num near relays\n")
-			return false
-		}
-
-		if entry.NumNearRelays > 0 {
-
-			for i := 0; i < int(entry.NumNearRelays); i++ {
-
-				if entry.NearRelayIDs[i] == 0 {
-					// Log this but do not return false
-					// TODO: investigate why nearRelayID is 0
-					fmt.Printf("NearRelayIDs[%d] is 0.\n%+v\n", i, entry)
-				}
-
-				if entry.NearRelayRTTs[i] < 0 || entry.NearRelayRTTs[i] > 255 {
-					fmt.Printf("invalid near relay rtt\n")
-					return false
-				}
-
-				if entry.NearRelayJitters[i] < 0 || entry.NearRelayJitters[i] > 255 {
-					fmt.Printf("invalid near relay jitter\n")
-					return false
-				}
-
-				if entry.NearRelayPacketLosses[i] < 0 || entry.NearRelayPacketLosses[i] > 100 {
-					fmt.Printf("invalid near relay packet loss\n")
-					return false
-				}
-			}
-		}
-	}
-
-	// network next only
-
-	if entry.Next {
-
-		if entry.NextRTT < 0 || entry.NextRTT > 255 {
-			fmt.Printf("invalid next rtt\n")
-			return false
-		}
-
-		if entry.NextJitter < 0 || entry.NextJitter > 255 {
-			fmt.Printf("invalid next jitter\n")
-			return false
-		}
-
-		if entry.NextPacketLoss < 0 || entry.NextPacketLoss > 100 {
-			fmt.Printf("invalid next packet loss\n")
-			return false
-		}
-
-		if entry.PredictedNextRTT < 0 || entry.PredictedNextRTT > 255 {
-			fmt.Printf("invalid predicted next rtt\n")
-			return false
-		}
-
-		if entry.NearRelayRTT < 0 || entry.NearRelayRTT > 255 {
-			fmt.Printf("invalid near relay rtt\n")
-			return false
-		}
-
-		if entry.NumNextRelays < 0 || entry.NumNextRelays > 32 {
-			fmt.Printf("invalid num next relays\n")
-			return false
-		}
-
-		if entry.RouteDiversity < 0 || entry.RouteDiversity > 31 {
-			fmt.Printf("invalid route diversity\n")
-			return false
-		}
-	}
-
-	return true
-}
-
-// Checks all floating point numbers in the BillingEntry2 for NaN and +-Inf and forces them to 0
-func (entry *BillingEntry2) CheckNaNOrInf() (bool, []string) {
-	var nanOrInfExists bool
-	var nanOrInfFields []string
-
-	if math.IsNaN(float64(entry.Latitude)) || math.IsInf(float64(entry.Latitude), 0) {
-		nanOrInfFields = append(nanOrInfFields, "Latitude")
-		nanOrInfExists = true
-		entry.Latitude = float32(0)
-	}
-
-	if math.IsNaN(float64(entry.Longitude)) || math.IsInf(float64(entry.Longitude), 0) {
-		nanOrInfFields = append(nanOrInfFields, "Longitude")
-		nanOrInfExists = true
-		entry.Longitude = float32(0)
-	}
-
-	return nanOrInfExists, nanOrInfFields
-}
-
-func (entry *BillingEntry2) Save() (map[string]bigquery.Value, string, error) {
-
-	e := make(map[string]bigquery.Value)
-
-	/*
-		1. Always
-
-		These values are written for every slice.
-	*/
-
-	e["timestamp"] = int(entry.Timestamp)
-	e["sessionID"] = int(entry.SessionID)
-	e["sliceNumber"] = int(entry.SliceNumber)
-	e["directRTT"] = int(entry.DirectRTT)
-	e["directJitter"] = int(entry.DirectJitter)
-	e["directPacketLoss"] = int(entry.DirectPacketLoss)
-	e["realPacketLoss"] = float64(entry.RealPacketLoss) + float64(entry.RealPacketLoss_Frac)/256.0
-	e["realJitter"] = int(entry.RealJitter)
-
-	if entry.Next {
-		e["next"] = true
-	}
-
-	if entry.Flagged {
-		e["flagged"] = true
-	}
-
-	if entry.Summary {
-		e["summary"] = true
-	}
-
-	if entry.UseDebug {
-		e["debug"] = entry.Debug
-	}
-
-	/*
-		2. First slice only
-
-		These values are serialized only for slice 0.
-	*/
-
-	if entry.SliceNumber == 0 {
-
-		e["datacenterID"] = int(entry.DatacenterID)
-		e["buyerID"] = int(entry.BuyerID)
-		e["userHash"] = int(entry.UserHash)
-		e["envelopeBytesUp"] = int(entry.EnvelopeBytesUp)
-		e["envelopeBytesDown"] = int(entry.EnvelopeBytesDown)
-		e["latitude"] = entry.Latitude
-		e["longitude"] = entry.Longitude
-		e["isp"] = entry.ISP
-		e["connectionType"] = int(entry.ConnectionType)
-		e["platformType"] = int(entry.PlatformType)
-		e["sdkVersion"] = entry.SDKVersion
-
-		if entry.NumTags > 0 {
-			tags := make([]bigquery.Value, entry.NumTags)
-			for i := 0; i < int(entry.NumTags); i++ {
-				tags[i] = int(entry.Tags[i])
-			}
-			e["tags"] = tags
-		}
-
-		if entry.ABTest {
-			e["abTest"] = true
-		}
-
-		if entry.Pro {
-			e["pro"] = true
-		}
-	}
-
-	/*
-		3. Summary slice only
-
-		These values are serialized only for the summary slice (at the end of the session)
-	*/
-
-	if entry.Summary {
-
-		e["clientToServerPacketsSent"] = int(entry.ClientToServerPacketsSent)
-		e["serverToClientPacketsSent"] = int(entry.ServerToClientPacketsSent)
-		e["clientToServerPacketsLost"] = int(entry.ClientToServerPacketsLost)
-		e["serverToClientPacketsLost"] = int(entry.ServerToClientPacketsLost)
-		e["clientToServerPacketsOutOfOrder"] = int(entry.ClientToServerPacketsOutOfOrder)
-		e["serverToClientPacketsOutOfOrder"] = int(entry.ServerToClientPacketsOutOfOrder)
-
-		if entry.NumNearRelays > 0 {
-
-			nearRelayIDs := make([]bigquery.Value, entry.NumNearRelays)
-			nearRelayRTTs := make([]bigquery.Value, entry.NumNearRelays)
-			nearRelayJitters := make([]bigquery.Value, entry.NumNearRelays)
-			nearRelayPacketLosses := make([]bigquery.Value, entry.NumNearRelays)
-
-			for i := 0; i < int(entry.NumNearRelays); i++ {
-				nearRelayIDs[i] = int(entry.NearRelayIDs[i])
-				nearRelayRTTs[i] = int(entry.NearRelayRTTs[i])
-				nearRelayJitters[i] = int(entry.NearRelayJitters[i])
-				nearRelayPacketLosses[i] = int(entry.NearRelayPacketLosses[i])
-			}
-
-			e["nearRelayIDs"] = nearRelayIDs
-			e["nearRelayRTTs"] = nearRelayRTTs
-			e["nearRelayJitters"] = nearRelayJitters
-			e["nearRelayPacketLosses"] = nearRelayPacketLosses
-
-		}
-
-	}
-
-	/*
-		4. Network Next Only
-
-		These values are serialized only when a slice is on network next.
-	*/
-
-	if entry.Next {
-
-		e["nextRTT"] = int(entry.NextRTT)
-		e["nextJitter"] = int(entry.NextJitter)
-		e["nextPacketLoss"] = int(entry.NextPacketLoss)
-		e["predictedNextRTT"] = int(entry.PredictedNextRTT)
-		e["nearRelayRTT"] = int(entry.NearRelayRTT)
-
-		if entry.NumNextRelays > 0 {
-
-			nextRelays := make([]bigquery.Value, entry.NumNextRelays)
-			nextRelayPrice := make([]bigquery.Value, entry.NumNextRelays)
-
-			for i := 0; i < int(entry.NumNextRelays); i++ {
-				nextRelays[i] = int(entry.NextRelays[i])
-				nextRelayPrice[i] = int(entry.NextRelayPrice[i])
-			}
-
-			e["nextRelays"] = nextRelays
-			e["nextRelayPrice"] = nextRelayPrice
-
-		}
-
-		e["totalPrice"] = int(entry.TotalPrice)
-		e["routeDiversity"] = int(entry.RouteDiversity)
-
-		if entry.Uncommitted {
-			e["uncommitted"] = true
-		}
-
-		if entry.Multipath {
-			e["multipath"] = true
-		}
-
-		if entry.RTTReduction {
-			e["rttReduction"] = true
-		}
-
-		if entry.PacketLossReduction {
-			e["packetLossReduction"] = true
-		}
-
-		if entry.RouteChanged {
-			e["routeChanged"] = true
-		}
-	}
-
-	/*
-		5. Error State Only
-
-		These values are only serialized when the session is in an error state (rare).
-	*/
-
-	if entry.FallbackToDirect {
-		e["fallbackToDirect"] = true
-	}
-
-	if entry.MultipathVetoed {
-		e["multipathVetoed"] = true
-	}
-
-	if entry.Mispredicted {
-		e["mispredicted"] = true
-	}
-
-	if entry.Vetoed {
-		e["vetoed"] = true
-	}
-
-	if entry.LatencyWorse {
-		e["latencyWorse"] = true
-	}
-
-	if entry.NoRoute {
-		e["noRoute"] = true
-	}
-
-	if entry.NextLatencyTooHigh {
-		e["nextLatencyTooHigh"] = true
-	}
-
-	if entry.CommitVeto {
-		e["commitVeto"] = true
-	}
-
-	if entry.UnknownDatacenter {
-		e["unknownDatacenter"] = true
-	}
-
-	if entry.DatacenterNotEnabled {
-		e["datacenterNotEnabled"] = true
-	}
-
-	if entry.BuyerNotLive {
-		e["buyerNotLive"] = true
-	}
-
-	if entry.StaleRouteMatrix {
-		e["staleRouteMatrix"] = true
-	}
-
-	return e, "", nil
 }
