+++ modules/billing/googlepubsub.go
@@ -37,7 +37,6 @@ type GooglePubSubClient struct {
 
 // NewBiller creates a new GooglePubSubBiller, sets up the pubsub clients, and starts goroutines to listen for publish results.
 // To clean up the results goroutine, use ctx.Done().
-// NOTE: use SEPARATE GooglePubSubBillers for writing BillingEntry and BillingEntry2 to utilize different pubsub topics
 func NewGooglePubSubBiller(ctx context.Context, billingMetrics *metrics.BillingMetrics, resultLogger log.Logger, projectID string, billingTopicID string, clientCount int, clientBufferCountThreshold int, clientMinBufferBytes int, settings *pubsub.PublishSettings) (Biller, error) {
 	if settings == nil {
 		return nil, errors.New("nil google pubsub publish settings")
@@ -119,53 +118,6 @@ func (biller *GooglePubSubBiller) Bill(ctx context.Context, entry *BillingEntry)
 		result = client.Topic.Publish(ctx, &pubsub.Message{Data: client.buffer})
 
 		client.Metrics.EntriesFlushed.Add(float64(client.bufferMessageCount))
-
-		client.buffer = make([]byte, 0)
-		client.bufferMessageCount = 0
-	}
-
-	client.bufferMutex.Unlock()
-
-	if result != nil {
-		client.ResultChan <- result
-	}
-
-	return nil
-}
-
-func (biller *GooglePubSubBiller) Bill2(ctx context.Context, entry *BillingEntry2) error {
-	if biller.clients == nil {
-		return fmt.Errorf("billing: clients not initialized")
-	}
-
-	index := entry.SessionID % uint64(len(biller.clients))
-	client := biller.clients[index]
-
-	entryBytes, err := WriteBillingEntry2(entry)
-	if err != nil {
-		return err
-	}
-
-	client.Metrics.PubsubBillingEntry2Size.Set(float64(len(entryBytes)))
-
-	data := make([]byte, 4+len(entryBytes))
-	var offset int
-	encoding.WriteUint32(data, &offset, uint32(len(entryBytes)))
-	encoding.WriteBytes(data, &offset, entryBytes, len(entryBytes))
-
-	client.bufferMutex.Lock()
-
-	if client.bufferMessageCount < client.BufferCountThreshold || len(client.buffer) < client.MinBufferBytes {
-		client.buffer = append(client.buffer, data...)
-		client.bufferMessageCount++
-		client.Metrics.Entries2Submitted.Add(1)
-	}
-
-	var result *pubsub.PublishResult
-	if client.bufferMessageCount >= client.BufferCountThreshold && len(client.buffer) >= client.MinBufferBytes {
-		result = client.Topic.Publish(ctx, &pubsub.Message{Data: client.buffer})
-
-		client.Metrics.Entries2Flushed.Add(float64(client.bufferMessageCount))
 
 		client.buffer = make([]byte, 0)
 		client.bufferMessageCount = 0
