+++ modules/billing/noop.go
@@ -16,12 +16,6 @@ func (noop *NoOpBiller) Bill(ctx context.Context, entry *BillingEntry) error {
 	return nil
 }
 
-// Bill2 does nothing
-func (noop *NoOpBiller) Bill2(ctx context.Context, entry *BillingEntry2) error {
-	atomic.AddUint64(&noop.submitted, 1)
-	return nil
-}
-
 // Close does nothing
 func (noop *NoOpBiller) Close() {}
