+++ modules/billing/pubsub_forwarder.go
@@ -23,8 +23,6 @@ type PubSubForwarder struct {
 	pubsubSubscription *pubsub.Subscription
 }
 
-// PubSubForwarder receives batches of billing entries from Google Pub/Sub, unbatches, and inserts them into an internal channel.
-// NOTE: use SEPARATE PubSubForwarders for receiving BillingEntry and BillingEntry2 from different Pub/Sub subscriptions.
 func NewPubSubForwarder(ctx context.Context, biller Biller, logger log.Logger, metrics *metrics.BillingMetrics, gcpProjectID string, topicName string, subscriptionName string, numRecvGoroutines int) (*PubSubForwarder, error) {
 	pubsubClient, err := pubsub.NewClient(ctx, gcpProjectID)
 	if err != nil {
@@ -101,65 +99,6 @@ func (psf *PubSubForwarder) Forward(ctx context.Context, wg *sync.WaitGroup) {
 				}
 
 				psf.Metrics.ErrorMetrics.BillingReadFailure.Add(1)
-				// Nack if we failed to read the billing entry
-				m.Nack()
-			}
-		}
-	})
-
-	if err != nil && err != context.Canceled {
-		// If the Receive function returns for any reason besides shutdown, we want to immediately exit and restart the service
-		level.Error(psf.Logger).Log("msg", "stopped receive loop", "err", err)
-		os.Exit(1)
-	}
-
-	// Close entries channel to ensure messages are drained for the final write to BigQuery
-	psf.Biller.Close()
-	level.Debug(psf.Logger).Log("msg", "receive canceled, closed entries channel")
-}
-
-// Forward reads the billing entry 2 from pubsub and writes it to BigQuery
-func (psf *PubSubForwarder) Forward2(ctx context.Context, wg *sync.WaitGroup) {
-	defer wg.Done()
-
-	err := psf.pubsubSubscription.Receive(ctx, func(ctx context.Context, m *pubsub.Message) {
-		entries, err := psf.unbatchMessages(m)
-		if err != nil {
-			level.Error(psf.Logger).Log("err", err)
-			psf.Metrics.ErrorMetrics.Billing2BatchedReadFailure.Add(1)
-		}
-
-		psf.Metrics.Entries2Received.Add(float64(len(entries)))
-
-		billingEntries := make([]BillingEntry2, len(entries))
-		for i := range billingEntries {
-			if err = ReadBillingEntry2(&billingEntries[i], entries[i]); err == nil {
-				if err := psf.Biller.Bill2(ctx, &billingEntries[i]); err != nil {
-					level.Error(psf.Logger).Log("msg", "could not submit billing entry", "err", err)
-					// Nack if we failed to submit the billing entry
-					m.Nack()
-					return
-				}
-
-				m.Ack()
-			} else {
-				entryVetoStr := os.Getenv("BILLING_ENTRY_VETO")
-				entryVeto, err := strconv.ParseBool(entryVetoStr)
-
-				if err != nil {
-					level.Error(psf.Logger).Log("msg", "failed to parse veto env var", "err", err)
-					psf.Metrics.ErrorMetrics.Billing2ReadFailure.Add(1)
-					// Nack if we failed to read the billing entry
-					m.Nack()
-					return
-				}
-
-				if entryVeto {
-					m.Ack()
-					return
-				}
-
-				psf.Metrics.ErrorMetrics.Billing2ReadFailure.Add(1)
 				// Nack if we failed to read the billing entry
 				m.Nack()
 			}
