+++ modules/core/core.go
@@ -1120,6 +1120,8 @@ func GetBestRoutes(routeMatrix []RouteEntry, sourceRelays []int32, sourceRelayCo
 			Debug("Number of routes in entry: %v", int(entry.NumRoutes))
 			for k := 0; k < int(entry.NumRoutes); k++ {
 				cost := entry.RouteCost[k] + sourceRelayCost[i]
+				Debug("route cost: %v", cost)
+				Debug("max cost: %v", cost)
 				if cost > maxCost {
 					break
 				}
@@ -1469,7 +1471,6 @@ type RouteShader struct {
 	BandwidthEnvelopeUpKbps   int32
 	BandwidthEnvelopeDownKbps int32
 	BannedUsers               map[uint64]bool
-	PacketLossSustained       float32
 }
 
 func NewRouteShader() RouteShader {
@@ -1488,7 +1489,6 @@ func NewRouteShader() RouteShader {
 		BandwidthEnvelopeUpKbps:   1024,
 		BandwidthEnvelopeDownKbps: 1024,
 		BannedUsers:               make(map[uint64]bool),
-		PacketLossSustained:       0,
 	}
 }
 
@@ -1532,7 +1532,6 @@ type RouteState struct {
 	MispredictCounter   uint32
 	LatencyWorseCounter uint32
 	MultipathRestricted bool
-	PLSustainedCounter  int32
 }
 
 type InternalConfig struct {
@@ -1689,7 +1688,7 @@ func MakeRouteDecision_TakeNetworkNext(routeMatrix []RouteEntry, routeShader *Ro
 	// should we try to reduce packet loss?
 
 	reducePacketLoss := false
-	if routeShader.ReducePacketLoss && ((directPacketLoss > routeShader.AcceptablePacketLoss && sliceNumber >= internal.ReducePacketLossMinSliceNumber) || routeState.PLSustainedCounter == 3) {
+	if routeShader.ReducePacketLoss && directPacketLoss > routeShader.AcceptablePacketLoss && sliceNumber >= internal.ReducePacketLossMinSliceNumber {
 		if debug != nil {
 			*debug += "try to reduce packet loss\n"
 		}
