+++ modules/core/core_test.go
@@ -3201,123 +3201,6 @@ func TestTakeNetworkNext_ReducePacketLoss_AfterMinSliceNumber(t *testing.T) {
 	assert.Equal(t, int32(1), test.routeDiversity)
 }
 
-func TestTakeNetworkNext_ReducePacketLoss_PLSustainedCounter_Not3(t *testing.T) {
-
-	t.Parallel()
-
-	env := NewTestEnvironment()
-
-	env.AddRelay("losangeles", "10.0.0.1")
-	env.AddRelay("chicago", "10.0.0.2")
-
-	env.SetCost("losangeles", "chicago", 10)
-
-	test := NewTestData(env)
-
-	test.directLatency = int32(20)
-	test.directPacketLoss = float32(5.0)
-
-	test.sourceRelays = []int32{0}
-	test.sourceRelayCosts = []int32{10}
-
-	test.destRelays = []int32{1}
-
-	test.routeShader.AcceptableLatency = 100
-	test.routeState.UserID = 100
-
-	test.routeShader.AcceptablePacketLoss = float32(20)
-	test.routeShader.PacketLossSustained = float32(2.0)
-
-	result := test.TakeNetworkNext()
-
-	assert.False(t, result)
-
-	expectedRouteState := RouteState{}
-	expectedRouteState.UserID = 100
-	expectedRouteState.Next = false
-	expectedRouteState.ReducePacketLoss = false
-	expectedRouteState.Committed = false
-
-	assert.Equal(t, expectedRouteState, test.routeState)
-	assert.Equal(t, int32(0), test.routeDiversity)
-
-	test.routeState.PLSustainedCounter = 1
-
-	result = test.TakeNetworkNext()
-
-	assert.False(t, result)
-
-	expectedRouteState = RouteState{}
-	expectedRouteState.UserID = 100
-	expectedRouteState.Next = false
-	expectedRouteState.ReducePacketLoss = false
-	expectedRouteState.Committed = false
-	expectedRouteState.PLSustainedCounter = 1
-
-	assert.Equal(t, expectedRouteState, test.routeState)
-	assert.Equal(t, int32(0), test.routeDiversity)
-
-	test.routeState.PLSustainedCounter = 2
-
-	result = test.TakeNetworkNext()
-
-	assert.False(t, result)
-
-	expectedRouteState = RouteState{}
-	expectedRouteState.UserID = 100
-	expectedRouteState.Next = false
-	expectedRouteState.ReducePacketLoss = false
-	expectedRouteState.Committed = false
-	expectedRouteState.PLSustainedCounter = 2
-
-	assert.Equal(t, expectedRouteState, test.routeState)
-	assert.Equal(t, int32(0), test.routeDiversity)
-}
-
-func TestTakeNetworkNext_ReducePacketLoss_PLSustainedCounter_3(t *testing.T) {
-
-	t.Parallel()
-
-	env := NewTestEnvironment()
-
-	env.AddRelay("losangeles", "10.0.0.1")
-	env.AddRelay("chicago", "10.0.0.2")
-
-	env.SetCost("losangeles", "chicago", 10)
-
-	test := NewTestData(env)
-
-	test.directLatency = int32(20)
-	test.directPacketLoss = float32(5.0)
-
-	test.sourceRelays = []int32{0}
-	test.sourceRelayCosts = []int32{10}
-
-	test.destRelays = []int32{1}
-
-	test.routeShader.AcceptableLatency = 100
-	test.routeState.UserID = 100
-
-	test.routeShader.AcceptablePacketLoss = float32(20)
-	test.routeShader.PacketLossSustained = float32(2.0)
-
-	test.routeState.PLSustainedCounter = 3
-
-	result := test.TakeNetworkNext()
-
-	assert.True(t, result)
-
-	expectedRouteState := RouteState{}
-	expectedRouteState.UserID = 100
-	expectedRouteState.Next = true
-	expectedRouteState.ReducePacketLoss = true
-	expectedRouteState.Committed = true
-	expectedRouteState.PLSustainedCounter = 3
-
-	assert.Equal(t, expectedRouteState, test.routeState)
-	assert.Equal(t, int32(1), test.routeDiversity)
-}
-
 // -----------------------------------------------------------------------------
 
 func TestTakeNetworkNext_ReduceLatency_Multipath(t *testing.T) {
