+++ modules/storage/new_sql.go
@@ -961,7 +961,6 @@ type sqlRouteShader struct {
 	ReducePacketLoss          bool
 	ReduceJitter              bool
 	SelectionPercent          int64
-	PacketLossSustained       float64
 }
 
 func (db *SQL) syncRouteShaders(ctx context.Context) error {
@@ -973,7 +972,7 @@ func (db *SQL) syncRouteShaders(ctx context.Context) error {
 
 	sql.Write([]byte("select ab_test, acceptable_latency, acceptable_packet_loss, bw_envelope_down_kbps, "))
 	sql.Write([]byte("bw_envelope_up_kbps, disable_network_next, latency_threshold, multipath, pro_mode, "))
-	sql.Write([]byte("reduce_latency, reduce_packet_loss, reduce_jitter, selection_percent, packet_loss_sustained, buyer_id from route_shaders "))
+	sql.Write([]byte("reduce_latency, reduce_packet_loss, reduce_jitter, selection_percent, buyer_id from route_shaders "))
 
 	rows, err := db.Client.QueryContext(ctx, sql.String())
 	if err != nil {
@@ -998,7 +997,6 @@ func (db *SQL) syncRouteShaders(ctx context.Context) error {
 			&sqlRS.ReducePacketLoss,
 			&sqlRS.ReduceJitter,
 			&sqlRS.SelectionPercent,
-			&sqlRS.PacketLossSustained,
 			&buyerID,
 		)
 		if err != nil {
@@ -1020,7 +1018,6 @@ func (db *SQL) syncRouteShaders(ctx context.Context) error {
 			AcceptablePacketLoss:      float32(sqlRS.AcceptablePacketLoss),
 			BandwidthEnvelopeUpKbps:   int32(sqlRS.BandwidthEnvelopeUpKbps),
 			BandwidthEnvelopeDownKbps: int32(sqlRS.BandwidthEnvelopeDownKbps),
-			PacketLossSustained:       float32(sqlRS.PacketLossSustained),
 		}
 
 		db.buyerMutex.RLock()
