+++ modules/storage/seed_sql_storage.go
@@ -425,7 +425,6 @@ func SeedSQLStorage(
 			ReduceLatency:             true,
 			ReducePacketLoss:          true,
 			SelectionPercent:          int(100),
-			PacketLossSustained:       0,
 		}
 
 		gaRouteShader := core.RouteShader{
