+++ modules/transport/packet_server.go
@@ -17,7 +17,7 @@ const (
 	MaxDatacenterNameLength = 256
 	MaxSessionUpdateRetries = 10
 
-	SessionDataVersion = 13
+	SessionDataVersion = 11
 
 	MaxSessionDataSize = 511
 
@@ -607,7 +607,6 @@ type SessionData struct {
 	PrevPacketsLostServerToClient uint64
 	HoldNearRelays                bool
 	HoldNearRelayRTT              [core.MaxNearRelays]int32
-	WroteSummary                  bool
 }
 
 func UnmarshalSessionData(sessionData *SessionData, data []byte) error {
@@ -768,20 +767,12 @@ func (sessionData *SessionData) Serialize(stream encoding.Stream) error {
 		}
 	}
 
-	if sessionData.Version >= 12 {
-		stream.SerializeInteger(&sessionData.RouteState.PLSustainedCounter, 0, 3)
-	}
-
-	if sessionData.Version >= 13 {
-		stream.SerializeBool(&sessionData.WroteSummary)
-	}
-
-	// IMPORTANT: ADD NEW FIELDS BELOW HERE ONLY.
+	// IMPORTANT: ADD NEW FIELDS BELOW HERE ONLY. 
 
 	// >>> new fields go here <<<
 
-	// IMPORTANT: ADD NEW FIELDS ABOVE HERE ONLY.
-	// AFTER YOU ADD NEW FIELDS, UPDATE THE COMMENTS SO FIELDS
+	// IMPORTANT: ADD NEW FIELDS ABOVE HERE ONLY. 
+	// AFTER YOU ADD NEW FIELDS, UPDATE THE COMMENTS SO FIELDS 
 	// MAY ONLY BE ADDED *AFTER* YOUR NEW FIELDS.
 	// FAILING TO FOLLOW THESE INSRUCTIONS WILL BREAK PRODUCTION!!!!
