+++ testdata/sqlite3-empty.sql
@@ -62,7 +62,6 @@ create table route_shaders (
   reduce_packet_loss boolean not null,
   reduce_jitter boolean not null,
   selection_percent integer not null,
-  packet_loss_sustained numeric not null,
   buyer_id integer not null unique,
   constraint fk_buyer_id foreign key (buyer_id) references buyers(id)
 );
@@ -155,7 +154,7 @@ create table metadata (
   sync_sequence_number bigint not null
 );
 
--- File generation: 2021/05/26 08:51:34
+-- File generation: 2021/05/22 12:54:51
 
 -- machine_types
 insert into machine_types values (0, 'none');
